Leverage sqlite3 module in production environment
In order for knex to work, sqlite3 must be installed in order to avoid a
lack of resolution when knex attempts to inject the database client
module.

Previously `npm install` alone, despite having sqlite3 in the
package.json file, did not work for reasons that are not yet clear.
This means that sqlite3 must be installed in the production environment
explicitly in order for the app to work.