Provide an interface for interacting with subscriber entity
The boundary will need to apply some business logic at some point to
validate email subscriptions, along with other requirements that are not
immediately evident. When this occurs, the `saveSubscriber` method will
become more robust.

As it's currently written, the request object and the object managed by
the business object will be coupled by the `req.body` property.